Removed CreateTable and StartGame Message, added StateMessages
These 2 messages are obsolote as there is only one table and the game
will start automatically if 4 players are at the table and ready.

Anyways 2 new Messages were added. The 4 players sitting at the table
can (before the game startet) change their state from ready to not ready
and so on... If all 4 players maintain ready for a short period of time
(or immediately) the server can start the game.